% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempconv.r
\name{convert.temp}
\alias{convert.temp}
\title{Converting Temperatures}
\usage{
convert.temp(from.temp, from.unit = "Celsius", to.unit = "Kelvin",
  as.json = FALSE)
}
\arguments{
\item{from.temp}{Temperature value to be converted.}

\item{from.unit}{Scale of the temperature value to be converted. Either
\code{"Celsius"} (default), \code{"Fahrenheit"} or \code{"Kelvin"}.}

\item{to.unit}{Target temperature scale. Either \code{"Celsius"} (default),
\code{"Fahrenheit"} or \code{"Kelvin"}.}

\item{as.json}{Indicates if function returns JSON object describing the
conversion instead of just the converted temperature value.}
}
\value{
Either the converted temperature value or a JSON object (see
  details).
}
\description{
Converts temperatures temperatures between degrees Celsius,
  degrees Fahrenheit and Kelvin.
}
\details{
If \code{as.json==TRUE} then \code{convert.temp()}  returns a JSON
  object containing two sub-object \code{from} and \code{to} each of which
  has a \code{temp} and a \code{unit} field to describe the temperature. The
  \code{from.unit} and \code{to.unit} arguments are case-insensitive.
}
\examples{
result.fahrh.json <- convert.temp(24.5, from.unit="celsius", to.unit="fahrenheit", as.json=TRUE)
result.kelv <- convert.temp(24.5)

}
