% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempconv.r
\name{tempconvertr}
\alias{tempconvertr}
\title{Package 'tempconvertr'}
\description{
Comfortable conversion of temperatures

 This package provides functionality to convert temperatures between degrees
 Celsius, degrees Fahrenheit and Kelvin in any direction. Its main function
 is \code{\link{convert.temp}()}.
}
